Split off `ParticipantRepository` from `StudyRepository`.
For now, `StudyRepository` still delegates to `ParticipantRepository`.
