+++ carp.studies.core/src/commonMain/kotlin/dk/cachet/carp/studies/domain/StudyRepository.kt
@@ -1,11 +1,14 @@
 package dk.cachet.carp.studies.domain
 
 import dk.cachet.carp.common.UUID
-import dk.cachet.carp.studies.domain.users.Participant
+import dk.cachet.carp.studies.domain.users.ParticipantRepository
 import dk.cachet.carp.studies.domain.users.StudyOwner
 
 
-interface StudyRepository
+/**
+ * Store [Study] instances.
+ */
+interface StudyRepository : ParticipantRepository
 {
     /**
      * Adds a new [study] to the repository.
@@ -30,19 +33,4 @@ interface StudyRepository
      * @throws IllegalArgumentException when no previous version of this study is stored in the repository.
      */
     suspend fun update( study: Study )
-
-    /**
-     * Adds a new [participant] for the study with [studyId] to the repository.
-     *
-     * @throws IllegalArgumentException when a study with the specified [studyId] does not exist,
-     * or when a participant with the specified ID already exists within the study.
-     */
-    suspend fun addParticipant( studyId: UUID, participant: Participant )
-
-    /**
-     * Returns the participants which were added to the study with the specified [studyId].
-     *
-     * @throws IllegalArgumentException when a study with the specified [studyId] does not exist.
-     */
-    suspend fun getParticipants( studyId: UUID ): List<Participant>
 }
