+++ carp.studies.core/src/commonTest/kotlin/dk/cachet/carp/studies/domain/StudyRepositoryTest.kt
@@ -1,9 +1,7 @@
 package dk.cachet.carp.studies.domain
 
 import dk.cachet.carp.common.UUID
-import dk.cachet.carp.common.users.AccountIdentity
 import dk.cachet.carp.deployment.domain.users.StudyInvitation
-import dk.cachet.carp.studies.domain.users.Participant
 import dk.cachet.carp.studies.domain.users.StudyOwner
 import dk.cachet.carp.test.runBlockingTest
 import kotlin.test.*
@@ -20,7 +18,8 @@ interface StudyRepositoryTest
     @Test
     fun cant_add_study_with_id_that_already_exists() = runBlockingTest {
         val repo = createRepository()
-        val study = addStudy( repo )
+        val study = Study( StudyOwner(), "Test" )
+        repo.add( study )
 
         val studyWithSameId = Study( StudyOwner(), "Study 2", "Description", StudyInvitation.empty(), study.id )
         assertFailsWith<IllegalArgumentException>
@@ -32,7 +31,8 @@ interface StudyRepositoryTest
     @Test
     fun getById_succeeds() = runBlockingTest {
         val repo = createRepository()
-        val study = addStudy( repo )
+        val study = Study( StudyOwner(), "Test" )
+        repo.add( study )
 
         val foundStudy = repo.getById( study.id )
         assertNotSame( study, foundStudy ) // Should be new object instance.
@@ -84,51 +84,5 @@ interface StudyRepositoryTest
 
         val study = Study( StudyOwner(), "Test" )
         assertFailsWith<IllegalArgumentException> { repo.update( study ) }
-    }
-
-    @Test
-    fun adding_participant_and_retrieving_it_succeeds() = runBlockingTest {
-        val repo = createRepository()
-        val study = addStudy( repo )
-
-        val participant = Participant( AccountIdentity.fromUsername( "user" ) )
-        repo.addParticipant( study.id, participant )
-        val studyParticipants = repo.getParticipants( study.id )
-        assertEquals( participant, studyParticipants.single() )
-    }
-
-    @Test
-    fun addParticipant_fails_with_nonexisting_studyId() = runBlockingTest {
-        val repo = createRepository()
-
-        val unknownId = UUID.randomUUID()
-        val participant = Participant( AccountIdentity.fromUsername( "user" ) )
-        assertFailsWith<IllegalArgumentException> { repo.addParticipant( unknownId, participant ) }
-    }
-
-    @Test
-    fun addParticipant_fails_for_duplicate_participant_id() = runBlockingTest {
-        val repo = createRepository()
-        val study = addStudy( repo )
-        val participant = Participant( AccountIdentity.fromUsername( "user" ) )
-        repo.addParticipant( study.id, participant )
-
-        assertFailsWith<IllegalArgumentException> { repo.addParticipant( study.id, participant ) }
-    }
-
-    @Test
-    fun getParticipants_fails_for_nonexisting_studyId() = runBlockingTest {
-        val repo = createRepository()
-
-        val unknownId = UUID.randomUUID()
-        assertFailsWith<IllegalArgumentException> { repo.getParticipants( unknownId ) }
-    }
-
-
-    private suspend fun addStudy( repo: StudyRepository ): Study
-    {
-        val study = Study( StudyOwner(), "Test")
-        repo.add( study )
-        return study
     }
 }
