+++ carp.studies.core/src/commonTest/kotlin/dk/cachet/carp/studies/domain/users/ParticipantRepositoryTest.kt
@@ -1,0 +1,37 @@
+package dk.cachet.carp.studies.domain.users
+
+import dk.cachet.carp.common.UUID
+import dk.cachet.carp.common.users.AccountIdentity
+import dk.cachet.carp.test.runBlockingTest
+import kotlin.test.*
+
+
+/**
+ * Tests for implementations of [ParticipantRepository].
+ */
+interface ParticipantRepositoryTest
+{
+    fun createRepository(): ParticipantRepository
+
+
+    @Test
+    fun adding_participant_and_retrieving_it_succeeds() = runBlockingTest {
+        val repo = createRepository()
+        val studyId = UUID.randomUUID()
+
+        val participant = Participant( AccountIdentity.fromUsername( "user" ) )
+        repo.addParticipant( studyId, participant )
+        val studyParticipants = repo.getParticipants( studyId )
+        assertEquals( participant, studyParticipants.single() )
+    }
+
+    @Test
+    fun addParticipant_fails_for_duplicate_participant_id() = runBlockingTest {
+        val repo = createRepository()
+        val studyId = UUID.randomUUID()
+        val participant = Participant( AccountIdentity.fromUsername( "user" ) )
+        repo.addParticipant( studyId, participant )
+
+        assertFailsWith<IllegalArgumentException> { repo.addParticipant( studyId, participant ) }
+    }
+}
